Lists saves sort/owner/timeCreated; Prepped Recipes/Activities for the same upgrade
The custom class implemented for Lists will have to be adapted for
Recipes/Activities be upgrade can be completed.

+++ app/src/main/java/com/aisleshare/ItemComparator.java
@@ -52,7 +52,7 @@ public  class ItemComparator{
     }
     public class Created implements Comparator<Item>{
         public int compare(Item left, Item right) {
-            double difference = left.getCreated() - right.getCreated();
+            long difference = left.getCreated() - right.getCreated();
             if(difference > 0){
                 return 1;
             }
